Basic Movement/Camera Setup
No real gameplay elements right now. Added some basic 3d movement and third person camera. The character model is slightly jittery while turning the camera currently. This game will most likely be a puzzle platformer. The theme is reflection so I'm thinking there will be a clone of the player that copies the same movements. Maybe the other clone has different properties or abilities than the player character that allows it to complete a section of the level? The player might also switch places with the other clone to gain its abilities in another section of the level. Nothing set in stone but just a few ideas.